#51: code cleanup: moved loader.io to file, made some spacing more consistent, divided functions up to routes/user_actions
