+++ views/welcome.erb
@@ -50,7 +50,7 @@
                   <%= t['text'] %>
                 </div>
               </div>
-              <% count = count + 1 %>
+              <% count = (count + 1)%100 %>
               <% end %>
 
             </div>
